+++ api/grpcserver/config.go
@@ -41,6 +41,7 @@ const (
 	ActivationStreamV2Alpha1 Service = "activation_stream_v2alpha1"
 	RewardV2Alpha1           Service = "reward_v2alpha1"
 	RewardStreamV2Alpha1     Service = "reward_stream_v2alpha1"
+	NetworkV2Alpha1          Service = "network_v2alpha1"
 )
 
 // DefaultConfig defines the default configuration options for api.
@@ -48,10 +49,13 @@ func DefaultConfig() Config {
 	return Config{
 		PublicServices: []Service{
 			GlobalState, Mesh, Transaction, Node, Activation, ActivationV2Alpha1,
-			RewardV2Alpha1,
+			NetworkV2Alpha1,
 		},
-		PublicListener:        "0.0.0.0:9092",
-		PrivateServices:       []Service{Admin, Smesher, Debug, ActivationStreamV2Alpha1, RewardStreamV2Alpha1},
+		PublicListener: "0.0.0.0:9092",
+		PrivateServices: []Service{
+			Admin, Smesher, Debug, ActivationStreamV2Alpha1,
+			RewardStreamV2Alpha1,
+		},
 		PrivateListener:       "127.0.0.1:9093",
 		PostServices:          []Service{Post, PostInfo},
 		PostListener:          "127.0.0.1:0",