+++ syncer/data_fetch_test.go
@@ -45,17 +45,6 @@ const (
 	numMalicious = 11
 )
 
-func generateMaliciousIDs(t *testing.T) ([]types.NodeID, []byte) {
-	t.Helper()
-	var malicious fetch.MaliciousIDs
-	for i := 0; i < numMalicious; i++ {
-		malicious.NodeIDs = append(malicious.NodeIDs, types.RandomNodeID())
-	}
-	data, err := codec.Encode(&malicious)
-	require.NoError(t, err)
-	return malicious.NodeIDs, data
-}
-
 func generateLayerOpinions(t *testing.T, bid *types.BlockID) []byte {
 	t.Helper()
 	lo := &fetch.LayerOpinion{
@@ -96,99 +85,6 @@ func GenPeers(num int) []p2p.Peer {
 		peers = append(peers, p2p.Peer(fmt.Sprintf("peer_%d", i)))
 	}
 	return peers
-}
-
-func TestDataFetch_PollMaliciousIDs(t *testing.T) {
-	numPeers := 4
-	peers := GenPeers(numPeers)
-	errUnknown := errors.New("unknown")
-	newTestDataFetchWithMocks := func(_ *testing.T, exists bool) *testDataFetch {
-		td := newTestDataFetch(t)
-		td.mFetcher.EXPECT().SelectBestShuffled(gomock.Any()).Return(peers)
-		for _, peer := range peers {
-			td.mFetcher.EXPECT().GetMaliciousIDs(gomock.Any(), peer).DoAndReturn(
-				func(_ context.Context, peer p2p.Peer) ([]byte, error) {
-					ids, data := generateMaliciousIDs(t)
-					for _, id := range ids {
-						td.mIDs.EXPECT().IdentityExists(id).Return(exists, nil)
-					}
-					return data, nil
-				})
-		}
-		return td
-	}
-	t.Run("getting malfeasance proofs success", func(t *testing.T) {
-		t.Parallel()
-		td := newTestDataFetchWithMocks(t, true)
-		td.mFetcher.EXPECT().GetMalfeasanceProofs(gomock.Any(), gomock.Any())
-		require.NoError(t, td.PollMaliciousProofs(context.Background()))
-	})
-	t.Run("getting proofs failure", func(t *testing.T) {
-		t.Parallel()
-		td := newTestDataFetchWithMocks(t, true)
-		td.mFetcher.EXPECT().GetMalfeasanceProofs(gomock.Any(), gomock.Any()).Return(errUnknown)
-		require.ErrorIs(t, td.PollMaliciousProofs(context.Background()), errUnknown)
-	})
-	t.Run("ids do not exist", func(t *testing.T) {
-		t.Parallel()
-		td := newTestDataFetchWithMocks(t, false)
-		td.mFetcher.EXPECT().GetMalfeasanceProofs(gomock.Any(), nil)
-		require.NoError(t, td.PollMaliciousProofs(context.Background()))
-	})
-}
-
-func TestDataFetch_PollMaliciousIDs_PeerErrors(t *testing.T) {
-	t.Run("malformed data in response", func(t *testing.T) {
-		t.Parallel()
-		peers := []p2p.Peer{"p0"}
-		td := newTestDataFetch(t)
-		td.mFetcher.EXPECT().SelectBestShuffled(gomock.Any()).Return(peers)
-		td.mFetcher.EXPECT().GetMaliciousIDs(gomock.Any(), p2p.Peer("p0")).Return([]byte("malformed"), nil)
-		err := td.PollMaliciousProofs(context.Background())
-		require.ErrorContains(t, err, "decode")
-	})
-	t.Run("peer fails", func(t *testing.T) {
-		t.Parallel()
-		peers := []p2p.Peer{"p0"}
-		expectedErr := errors.New("peer failure")
-		td := newTestDataFetch(t)
-		td.mFetcher.EXPECT().SelectBestShuffled(gomock.Any()).Return(peers)
-		td.mFetcher.EXPECT().GetMaliciousIDs(gomock.Any(), p2p.Peer("p0")).Return(nil, expectedErr)
-		err := td.PollMaliciousProofs(context.Background())
-		require.ErrorIs(t, err, expectedErr)
-	})
-	t.Run("one peer sends malformed data (succeed anyway)", func(t *testing.T) {
-		t.Parallel()
-		peers := []p2p.Peer{"p0", "p1"}
-		td := newTestDataFetch(t)
-		maliciousIds, data := generateMaliciousIDs(t)
-		for _, id := range maliciousIds {
-			td.mIDs.EXPECT().IdentityExists(id).Return(true, nil)
-		}
-
-		td.mFetcher.EXPECT().SelectBestShuffled(gomock.Any()).Return(peers)
-		td.mFetcher.EXPECT().GetMaliciousIDs(gomock.Any(), p2p.Peer("p0")).Return(data, nil)
-		td.mFetcher.EXPECT().GetMaliciousIDs(gomock.Any(), p2p.Peer("p1")).Return([]byte("malformed"), nil)
-		td.mFetcher.EXPECT().GetMalfeasanceProofs(gomock.Any(), gomock.Any())
-		err := td.PollMaliciousProofs(context.Background())
-		require.NoError(t, err)
-	})
-	t.Run("one peer fails (succeed anyway)", func(t *testing.T) {
-		t.Parallel()
-		peers := []p2p.Peer{"p0", "p1"}
-		expectedErr := errors.New("peer failure")
-		td := newTestDataFetch(t)
-		maliciousIds, data := generateMaliciousIDs(t)
-		for _, id := range maliciousIds {
-			td.mIDs.EXPECT().IdentityExists(id).Return(true, nil)
-		}
-		td.mFetcher.EXPECT().SelectBestShuffled(gomock.Any()).Return(peers)
-		td.mFetcher.EXPECT().GetMaliciousIDs(gomock.Any(), p2p.Peer("p0")).Return(data, nil)
-		td.mFetcher.EXPECT().GetMaliciousIDs(gomock.Any(), p2p.Peer("p1")).Return(nil, expectedErr)
-		td.mFetcher.EXPECT().GetMalfeasanceProofs(gomock.Any(), gomock.Any())
-		err := td.PollMaliciousProofs(context.Background())
-		require.NoError(t, err)
-	})
 }
 
 func TestDataFetch_PollLayerData(t *testing.T) {