+++ syncer/syncer.go
@@ -13,10 +13,12 @@ import (
 	"github.com/spacemeshos/go-spacemesh/common/types"
 	"github.com/spacemeshos/go-spacemesh/datastore"
 	"github.com/spacemeshos/go-spacemesh/events"
+	"github.com/spacemeshos/go-spacemesh/fetch"
 	"github.com/spacemeshos/go-spacemesh/log"
 	"github.com/spacemeshos/go-spacemesh/mesh"
 	"github.com/spacemeshos/go-spacemesh/p2p"
 	"github.com/spacemeshos/go-spacemesh/syncer/atxsync"
+	"github.com/spacemeshos/go-spacemesh/syncer/malsync"
 	"github.com/spacemeshos/go-spacemesh/system"
 )
 
@@ -35,6 +37,7 @@ type Config struct {
 	DisableMeshAgreement     bool           `mapstructure:"disable-mesh-agreement"`
 	OutOfSyncThresholdLayers uint32         `mapstructure:"out-of-sync-threshold"`
 	AtxSync                  atxsync.Config `mapstructure:"atx-sync"`
+	MalSync                  malsync.Config `mapstructure:"malfeasance-sync"`
 }
 
 // DefaultConfig for the syncer.
@@ -49,6 +52,7 @@ func DefaultConfig() Config {
 		GossipDuration:           15 * time.Second,
 		OutOfSyncThresholdLayers: 3,
 		AtxSync:                  atxsync.DefaultConfig(),
+		MalSync:                  malsync.DefaultConfig(),
 	}
 }
 
@@ -121,6 +125,7 @@ type Syncer struct {
 	cfg          Config
 	cdb          *datastore.CachedDB
 	atxsyncer    atxSyncer
+	malsyncer    malSyncer
 	ticker       layerTicker
 	beacon       system.BeaconGetter
 	mesh         *mesh.Mesh
@@ -146,6 +151,12 @@ type Syncer struct {
 		cancel context.CancelFunc
 	}
 
+	// malSync runs malfeasant identity sync in the background
+	malSync struct {
+		started bool
+		eg      errgroup.Group
+	}
+
 	// awaitATXSyncedCh is the list of subscribers' channels to notify when this node enters ATX synced state
 	awaitATXSyncedCh chan struct{}
 
@@ -164,6 +175,7 @@ func NewSyncer(
 	patrol layerPatrol,
 	ch certHandler,
 	atxSyncer atxSyncer,
+	malSyncer malSyncer,
 	opts ...Option,
 ) *Syncer {
 	s := &Syncer{
@@ -171,6 +183,7 @@ func NewSyncer(
 		cfg:              DefaultConfig(),
 		cdb:              cdb,
 		atxsyncer:        atxSyncer,
+		malsyncer:        malSyncer,
 		ticker:           ticker,
 		beacon:           beacon,
 		mesh:             mesh,
@@ -405,12 +418,18 @@ func (s *Syncer) synchronize(ctx context.Context) bool {
 		// always sync to currentLayer-1 to reduce race with gossip and hare/tortoise
 		for layer := s.getLastSyncedLayer().Add(1); layer.Before(s.ticker.CurrentLayer()); layer = layer.Add(1) {
 			if err := s.syncLayer(ctx, layer); err != nil {
-				if !errors.Is(err, context.Canceled) {
-					// BatchError spams too much, in case of no progress enable debug mode for sync
+				batchError := &fetch.BatchError{}
+				if errors.As(err, &batchError) && batchError.Ignore() {
 					s.logger.With().
-						Debug("failed to sync layer", log.Context(ctx), log.Err(err), layer)
+						Info("remaining ballots are rejected in the layer", log.Context(ctx), log.Err(err), layer)
+				} else {
+					if !errors.Is(err, context.Canceled) {
+						// BatchError spams too much, in case of no progress enable debug mode for sync
+						s.logger.With().
+							Debug("failed to sync layer", log.Context(ctx), log.Err(err), layer)
+					}
+					return false
 				}
-				return false
 			}
 			s.setLastSyncedLayer(layer)
 		}
@@ -449,7 +468,7 @@ func (s *Syncer) syncAtx(ctx context.Context) error {
 
 		// FIXME https://github.com/spacemeshos/go-spacemesh/issues/3987
 		s.logger.With().Info("syncing malicious proofs", log.Context(ctx))
-		if err := s.syncMalfeasance(ctx); err != nil {
+		if err := s.syncMalfeasance(ctx, current.GetEpoch()); err != nil {
 			return err
 		}
 		s.logger.With().Info("malicious IDs synced", log.Context(ctx))
@@ -458,6 +477,7 @@ func (s *Syncer) syncAtx(ctx context.Context) error {
 
 	publish := current.GetEpoch()
 	if publish == 0 {
+		s.logger.With().Info("QQQQQ: nothing to sync")
 		return nil // nothing to sync in epoch 0
 	}
 
@@ -488,6 +508,22 @@ func (s *Syncer) syncAtx(ctx context.Context) error {
 			}
 			s.backgroundSync.epoch.Store(0)
 			return err
+		})
+	}
+	s.logger.With().Info("QQQQQ: aaaa")
+	if !s.malSync.started {
+		s.malSync.started = true
+		s.malSync.eg.Go(func() error {
+			select {
+			case <-ctx.Done():
+				return nil
+			case <-s.awaitATXSyncedCh:
+				err := s.malsyncer.DownloadLoop(ctx)
+				if err != nil && !errors.Is(err, context.Canceled) {
+					s.logger.WithContext(ctx).Error("malfeasance sync failed", log.Err(err))
+				}
+				return nil
+			}
 		})
 	}
 	return nil
@@ -578,16 +614,18 @@ func (s *Syncer) setStateAfterSync(ctx context.Context, success bool) {
 	}
 }
 
-func (s *Syncer) syncMalfeasance(ctx context.Context) error {
-	if err := s.dataFetcher.PollMaliciousProofs(ctx); err != nil {
-		return fmt.Errorf("PollMaliciousProofs: %w", err)
+func (s *Syncer) syncMalfeasance(ctx context.Context, epoch types.EpochID) error {
+	epochStart := s.ticker.LayerToTime(epoch.FirstLayer())
+	epochEnd := s.ticker.LayerToTime(epoch.Add(1).FirstLayer())
+	if err := s.malsyncer.EnsureInSync(ctx, epochStart, epochEnd); err != nil {
+		return fmt.Errorf("syncing malfeasance proof: %w", err)
 	}
 	return nil
 }
 
 func (s *Syncer) syncLayer(ctx context.Context, layerID types.LayerID, peers ...p2p.Peer) error {
 	if err := s.dataFetcher.PollLayerData(ctx, layerID, peers...); err != nil {
-		return fmt.Errorf("download layer data %v: %w", layerID, err)
+		return err
 	}
 	dataLayer.Set(float64(layerID))
 	return nil