+++ syncer/syncer_test.go
@@ -65,6 +65,7 @@ func (mlt *mockLayerTicker) LayerToTime(layerID types.LayerID) time.Time {
 }
 
 type testSyncer struct {
+	t       testing.TB
 	syncer  *Syncer
 	cdb     *datastore.CachedDB
 	msh     *mesh.Mesh
@@ -72,6 +73,7 @@ type testSyncer struct {
 
 	mDataFetcher *mocks.MockfetchLogic
 	mAtxSyncer   *mocks.MockatxSyncer
+	mMalSyncer   *mocks.MockmalSyncer
 	mBeacon      *smocks.MockBeaconGetter
 	mLyrPatrol   *mocks.MocklayerPatrol
 	mVm          *mmocks.MockvmState
@@ -81,15 +83,42 @@ type testSyncer struct {
 	mForkFinder  *mocks.MockforkFinder
 }
 
+func (ts *testSyncer) expectMalEnsureInSync(current types.LayerID) {
+	ts.mMalSyncer.EXPECT().EnsureInSync(
+		gomock.Any(),
+		ts.mTicker.LayerToTime(current.GetEpoch().FirstLayer()),
+		ts.mTicker.LayerToTime(current.GetEpoch().Add(1).FirstLayer()),
+	)
+}
+
+func (ts *testSyncer) expectDownloadLoop() chan struct{} {
+	ch := make(chan struct{})
+	ts.mMalSyncer.EXPECT().DownloadLoop(gomock.Any()).
+		DoAndReturn(func(context.Context) error {
+			close(ch)
+			return nil
+		})
+	ts.t.Cleanup(func() {
+		select {
+		case <-ch:
+		case <-time.After(10 * time.Second):
+			require.FailNow(ts.t, "timed out waiting for malsync loop start")
+		}
+	})
+	return ch
+}
+
 func newTestSyncer(t *testing.T, interval time.Duration) *testSyncer {
 	lg := logtest.New(t)
 	mt := newMockLayerTicker()
 	ctrl := gomock.NewController(t)
 
 	ts := &testSyncer{
+		t:            t,
 		mTicker:      mt,
 		mDataFetcher: mocks.NewMockfetchLogic(ctrl),
 		mAtxSyncer:   mocks.NewMockatxSyncer(ctrl),
+		mMalSyncer:   mocks.NewMockmalSyncer(ctrl),
 		mBeacon:      smocks.NewMockBeaconGetter(ctrl),
 		mLyrPatrol:   mocks.NewMocklayerPatrol(ctrl),
 		mVm:          mmocks.NewMockvmState(ctrl),
@@ -124,6 +153,7 @@ func newTestSyncer(t *testing.T, interval time.Duration) *testSyncer {
 		ts.mLyrPatrol,
 		ts.mCertHdr,
 		ts.mAtxSyncer,
+		ts.mMalSyncer,
 		WithConfig(cfg),
 		WithLogger(lg),
 		withDataFetcher(ts.mDataFetcher),
@@ -173,10 +203,11 @@ func TestSynchronize_OnlyOneSynchronize(t *testing.T) {
 	ts.mTicker.advanceToLayer(current)
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
+	dlCh := ts.expectDownloadLoop()
 	ts.syncer.Start()
 
 	ts.mAtxSyncer.EXPECT().Download(gomock.Any(), gomock.Any(), gomock.Any()).AnyTimes()
-	ts.mDataFetcher.EXPECT().PollMaliciousProofs(gomock.Any())
+	ts.expectMalEnsureInSync(current)
 	gLayer := types.GetEffectiveGenesis()
 
 	started := make(chan struct{}, 1)
@@ -203,6 +234,7 @@ func TestSynchronize_OnlyOneSynchronize(t *testing.T) {
 	// allow synchronize to finish
 	close(done)
 	require.NoError(t, eg.Wait())
+	<-dlCh
 
 	cancel()
 	ts.syncer.Close()
@@ -232,6 +264,7 @@ func advanceState(t testing.TB, ts *testSyncer, from, to types.LayerID) {
 
 func TestSynchronize_AllGood(t *testing.T) {
 	ts := newSyncerWithoutPeriodicRuns(t)
+	ts.expectDownloadLoop()
 	gLayer := types.GetEffectiveGenesis()
 	current1 := gLayer.Add(10)
 	ts.mTicker.advanceToLayer(current1)
@@ -257,7 +290,7 @@ func TestSynchronize_AllGood(t *testing.T) {
 			})
 	}
 
-	ts.mDataFetcher.EXPECT().PollMaliciousProofs(gomock.Any())
+	ts.expectMalEnsureInSync(current1)
 	for lid := gLayer.Add(1); lid.Before(current2); lid = lid.Add(1) {
 		ts.mDataFetcher.EXPECT().PollLayerData(gomock.Any(), lid)
 	}
@@ -306,13 +339,14 @@ func TestSynchronize_AllGood(t *testing.T) {
 
 func TestSynchronize_FetchLayerDataFailed(t *testing.T) {
 	ts := newSyncerWithoutPeriodicRuns(t)
+	ts.expectDownloadLoop()
 	gLayer := types.GetEffectiveGenesis()
 	current := gLayer.Add(2)
 	ts.mTicker.advanceToLayer(current)
 	lyr := current.Sub(1)
 	// times 2 as we will also spinup background worker
 	ts.mAtxSyncer.EXPECT().Download(gomock.Any(), gLayer.GetEpoch(), gomock.Any()).Times(2)
-	ts.mDataFetcher.EXPECT().PollMaliciousProofs(gomock.Any())
+	ts.expectMalEnsureInSync(current)
 	ts.mDataFetcher.EXPECT().PollLayerData(gomock.Any(), lyr).Return(errors.New("meh"))
 
 	require.False(t, ts.syncer.synchronize(context.Background()))
@@ -332,7 +366,7 @@ func TestSynchronize_FetchMalfeasanceFailed(t *testing.T) {
 	ts.mTicker.advanceToLayer(current)
 	lyr := current.Sub(1)
 	ts.mAtxSyncer.EXPECT().Download(gomock.Any(), gomock.Any(), gomock.Any()).AnyTimes()
-	ts.mDataFetcher.EXPECT().PollMaliciousProofs(gomock.Any()).Return(errors.New("meh"))
+	ts.mMalSyncer.EXPECT().EnsureInSync(gomock.Any(), gomock.Any(), gomock.Any()).Return(errors.New("meh"))
 
 	require.False(t, ts.syncer.synchronize(context.Background()))
 	require.EqualValues(t, current.GetEpoch()-1, ts.syncer.lastAtxEpoch())
@@ -388,7 +422,7 @@ func startWithSyncedState(t *testing.T, ts *testSyncer) types.LayerID {
 
 	gLayer := types.GetEffectiveGenesis()
 	ts.mTicker.advanceToLayer(gLayer)
-	ts.mDataFetcher.EXPECT().PollMaliciousProofs(gomock.Any())
+	ts.expectMalEnsureInSync(gLayer)
 	ts.mAtxSyncer.EXPECT().Download(gomock.Any(), gLayer.GetEpoch(), gomock.Any())
 	require.True(t, ts.syncer.synchronize(context.Background()))
 	ts.syncer.waitBackgroundSync()
@@ -418,8 +452,10 @@ func TestSyncAtxs_Genesis(t *testing.T) {
 	})
 	t.Run("first atx epoch", func(t *testing.T) {
 		ts := newSyncerWithoutPeriodicRuns(t)
+		ts.expectDownloadLoop()
 		epoch := types.EpochID(1)
-		ts.mTicker.advanceToLayer(epoch.FirstLayer() + 2) // to pass epoch end fraction threshold
+		current := epoch.FirstLayer() + 2
+		ts.mTicker.advanceToLayer(current) // to pass epoch end fraction threshold
 		require.False(t, ts.syncer.ListenToATXGossip())
 		wait := make(chan types.EpochID, 1)
 		ts.mAtxSyncer.EXPECT().
@@ -432,7 +468,7 @@ func TestSyncAtxs_Genesis(t *testing.T) {
 				}
 				return nil
 			})
-		ts.mDataFetcher.EXPECT().PollMaliciousProofs(gomock.Any())
+		ts.expectMalEnsureInSync(current)
 		require.True(t, ts.syncer.synchronize(context.Background()))
 		require.True(t, ts.syncer.ListenToATXGossip())
 		select {
@@ -465,6 +501,7 @@ func TestSyncAtxs(t *testing.T) {
 		tc := tc
 		t.Run(tc.desc, func(t *testing.T) {
 			ts := newSyncerWithoutPeriodicRuns(t)
+			ts.expectDownloadLoop()
 			lyr := startWithSyncedState(t, ts)
 			require.LessOrEqual(t, lyr, tc.current)
 			ts.mTicker.advanceToLayer(tc.current)
@@ -482,6 +519,7 @@ func TestSyncAtxs(t *testing.T) {
 
 func TestSynchronize_StaySyncedUponFailure(t *testing.T) {
 	ts := newSyncerWithoutPeriodicRuns(t)
+	ts.expectDownloadLoop()
 	lyr := startWithSyncedState(t, ts)
 	current := lyr.Add(1)
 	ts.mTicker.advanceToLayer(current)
@@ -498,6 +536,7 @@ func TestSynchronize_StaySyncedUponFailure(t *testing.T) {
 
 func TestSynchronize_BecomeNotSyncedUponFailureIfNoGossip(t *testing.T) {
 	ts := newSyncerWithoutPeriodicRuns(t)
+	ts.expectDownloadLoop()
 	lyr := startWithSyncedState(t, ts)
 	current := lyr.Add(outOfSyncThreshold)
 	ts.mTicker.advanceToLayer(current)
@@ -516,12 +555,13 @@ func TestSynchronize_BecomeNotSyncedUponFailureIfNoGossip(t *testing.T) {
 // test the case where the node originally starts from notSynced and eventually becomes synced.
 func TestFromNotSyncedToSynced(t *testing.T) {
 	ts := newSyncerWithoutPeriodicRuns(t)
+	ts.expectDownloadLoop()
 	ts.mAtxSyncer.EXPECT().Download(gomock.Any(), gomock.Any(), gomock.Any()).AnyTimes()
-	ts.mDataFetcher.EXPECT().PollMaliciousProofs(gomock.Any())
 	lyr := types.GetEffectiveGenesis().Add(1)
 	current := lyr.Add(5)
 	ts.mTicker.advanceToLayer(current)
 	ts.mDataFetcher.EXPECT().PollLayerData(gomock.Any(), lyr).Return(errors.New("baa-ram-ewe"))
+	ts.expectMalEnsureInSync(current)
 
 	require.False(t, ts.syncer.synchronize(context.Background()))
 	require.False(t, ts.syncer.dataSynced())
@@ -550,11 +590,12 @@ func TestFromNotSyncedToSynced(t *testing.T) {
 // to notSynced.
 func TestFromGossipSyncToNotSynced(t *testing.T) {
 	ts := newSyncerWithoutPeriodicRuns(t)
+	ts.expectDownloadLoop()
 	ts.mAtxSyncer.EXPECT().Download(gomock.Any(), gomock.Any(), gomock.Any()).AnyTimes()
 	lyr := types.GetEffectiveGenesis().Add(1)
 	current := lyr.Add(1)
 	ts.mTicker.advanceToLayer(current)
-	ts.mDataFetcher.EXPECT().PollMaliciousProofs(gomock.Any())
+	ts.expectMalEnsureInSync(current)
 	ts.mDataFetcher.EXPECT().PollLayerData(gomock.Any(), lyr)
 
 	require.True(t, ts.syncer.synchronize(context.Background()))
@@ -581,6 +622,7 @@ func TestFromGossipSyncToNotSynced(t *testing.T) {
 
 func TestNetworkHasNoData(t *testing.T) {
 	ts := newSyncerWithoutPeriodicRuns(t)
+	ts.expectDownloadLoop()
 	lyr := startWithSyncedState(t, ts)
 	require.True(t, ts.syncer.IsSynced(context.Background()))
 
@@ -606,8 +648,8 @@ func TestNetworkHasNoData(t *testing.T) {
 // eventually become synced again.
 func TestFromSyncedToNotSynced(t *testing.T) {
 	ts := newSyncerWithoutPeriodicRuns(t)
+	ts.expectDownloadLoop()
 	ts.mAtxSyncer.EXPECT().Download(gomock.Any(), gomock.Any(), gomock.Any()).AnyTimes()
-	ts.mDataFetcher.EXPECT().PollMaliciousProofs(gomock.Any()).AnyTimes()
 
 	require.True(t, ts.syncer.synchronize(context.Background()))
 	require.True(t, ts.syncer.IsSynced(context.Background()))
@@ -657,11 +699,12 @@ func waitOutGossipSync(t *testing.T, ts *testSyncer) {
 func TestSync_AlsoSyncProcessedLayer(t *testing.T) {
 	ts := newSyncerWithoutPeriodicRuns(t)
 
+	ts.expectDownloadLoop()
 	ts.mAtxSyncer.EXPECT().Download(gomock.Any(), gomock.Any(), gomock.Any()).AnyTimes()
-	ts.mDataFetcher.EXPECT().PollMaliciousProofs(gomock.Any())
 	lyr := types.GetEffectiveGenesis().Add(1)
 	current := lyr.Add(1)
 	ts.mTicker.advanceToLayer(current)
+	ts.expectMalEnsureInSync(current)
 
 	// simulate hare advancing the mesh forward
 	ts.mTortoise.EXPECT().TallyVotes(gomock.Any(), lyr)
@@ -717,6 +760,7 @@ func TestSyncer_IsBeaconSynced(t *testing.T) {
 
 func TestSynchronize_RecoverFromCheckpoint(t *testing.T) {
 	ts := newSyncerWithoutPeriodicRuns(t)
+	ts.expectDownloadLoop()
 	current := types.GetEffectiveGenesis().Add(types.GetLayersPerEpoch() * 5)
 	// recover from a checkpoint
 	types.SetEffectiveGenesis(current.Uint32())
@@ -731,6 +775,7 @@ func TestSynchronize_RecoverFromCheckpoint(t *testing.T) {
 		ts.mLyrPatrol,
 		ts.mCertHdr,
 		ts.mAtxSyncer,
+		ts.mMalSyncer,
 		WithConfig(ts.syncer.cfg),
 		WithLogger(ts.syncer.logger),
 		withDataFetcher(ts.mDataFetcher),
@@ -739,7 +784,7 @@ func TestSynchronize_RecoverFromCheckpoint(t *testing.T) {
 	// should not sync any atxs before current epoch
 	ts.mAtxSyncer.EXPECT().Download(gomock.Any(), current.GetEpoch(), gomock.Any())
 
-	ts.mDataFetcher.EXPECT().PollMaliciousProofs(gomock.Any())
+	ts.expectMalEnsureInSync(current)
 	require.True(t, ts.syncer.synchronize(context.Background()))
 	ts.syncer.waitBackgroundSync()
 	require.Equal(t, current.GetEpoch(), ts.syncer.lastAtxEpoch())